Move all the macros into a single kremlib file
